demo/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==24.1.1
certifi==2025.7.9
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.4
distlib==0.4.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.2
filelock==3.19.1
flake8==7.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.13
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.11.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.19.2
pytest==7.4.4
pytest-asyncio==0.23.0
pytest-cov==4.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
typer==0.16.0
types-PyYAML==6.0.12.20250809
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
zipp==3.23.0
openai>=1.0.0


 --extra-index-url https://pypi.org/simple/
 markdown-flow==0.2.35
# --extra-index-url https://test.pypi.org/simple/
# markdown-flow==0.2.34a3
